1029 Improve flow of login when a user clicks a team invitation link.
Case 1: User is not logged in:
Taken to /login with a get param team_invitation that is just /:ngo/:ngo_chal/add_team_member/:token and forward it to
Strava's redirect_uri so that when the user is done, he's redirected to the add_team_member endpoint.

Case 2: User is already logged in:
User will be directly take to add_team_member endpoint and added to team.